ubbink-server/requirements.txt
fastapi
uvicorn
python-dotenv
pymodbus==2.5.3
pyubbink
